added rotation
the line can be rotated using the 'j' and 'k' keys.
Unkike in the previous project, this time the rotation is
correct and the line does not 'gravitate' towards the
lower-left corner of the window. That behaviour due to
using ints to store coordinates. The new classes ( such
as xmx_Line ) store their coordinates in GLfloat's.