giving up on dynamic allocation for Bezier
The bug proved a tough one, couldn't fix it. It looks that
it's related to how the way glMap1f interprets the pointer given
as the 6th parameter.

Instead, I decided to have a BezierCurve class - stores 4 points
( 2 end points, 2 control points ) and, later, add a BezierPath
class, which will be an aggregate of BezierCurve objects.