working on rotating a Bezier
It looks like using the feedback buffer is not going to work for
capturing the transformed vertices: whenever, after rotating, a
point falls outside the visible area, it is clipped. There probably
is some way of achieving this ( making the viewing area bigger,
temprary ? ) but it looks easier to just compute the rotations for
the 4 points 'by hand' - which is what I'll try next.